fix pom
