+++ dubbo/src/main/resources/dubbo-business.xml
@@ -24,10 +24,10 @@
     <dubbo:application name="dubbo-demo-app"  />
     <dubbo:registry address="multicast://224.5.6.7:1234?unicast=false" />
 
-    <dubbo:reference id="orderService" check="false" interface="com.alibaba.fescar.tm.dubbo.OrderService"/>
-    <dubbo:reference id="storageService" check="false" interface="com.alibaba.fescar.tm.dubbo.StorageService"/>
+    <dubbo:reference id="orderService" check="false" interface="com.alibaba.fescar.samples.dubbo.OrderService"/>
+    <dubbo:reference id="storageService" check="false" interface="com.alibaba.fescar.samples.dubbo.StorageService"/>
 
-    <bean id="business" class="com.alibaba.fescar.tm.dubbo.impl.BusinessServiceImpl">
+    <bean id="business" class="com.alibaba.fescar.samples.dubbo.impl.BusinessServiceImpl">
         <property name="orderService" ref="orderService"/>
         <property name="storageService" ref="storageService"/>
     </bean>
