+++ dubbo/src/main/resources/dubbo-storage-service.xml
@@ -61,9 +61,9 @@
     <!--support nacos-->
     <!--<dubbo:registry address="nacos://127.0.0.1:8848"/>-->
     <dubbo:protocol name="dubbo" port="20882" />
-    <dubbo:service interface="com.alibaba.fescar.tm.dubbo.StorageService" ref="service" timeout="10000"/>
+    <dubbo:service interface="com.alibaba.fescar.samples.dubbo.StorageService" ref="service" timeout="10000"/>
 
-    <bean id="service" class="com.alibaba.fescar.tm.dubbo.impl.StorageServiceImpl">
+    <bean id="service" class="com.alibaba.fescar.samples.dubbo.impl.StorageServiceImpl">
         <property name="jdbcTemplate" ref="jdbcTemplate"/>
     </bean>
