+++ pom.xml
@@ -22,7 +22,7 @@
     <artifactId>fescar-samples</artifactId>
     <packaging>pom</packaging>
     <version>0.1.4-SNAPSHOT</version>
-    <name>fescar-samples ${fescar.version}</name>
+    <name>fescar-samples ${project.version}</name>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -57,7 +57,6 @@
     <modules>
         <module>dubbo</module>
         <module>springboot</module>
-        <module>common</module>
     </modules>
     <dependencyManagement>
         <dependencies>
