+++ springboot/pom.xml
@@ -28,7 +28,7 @@
 		<spring.version>5.0.8.RELEASE</spring.version>
 		<springboot.version>2.0.4.RELEASE</springboot.version>
 		<junit.version>4.12</junit.version>
-		<fescar.version>0.1.2-SNAPSHOT</fescar.version>
+		<fescar.version>0.1.3</fescar.version>
 	</properties>
 	<dependencyManagement>
 		<dependencies>
